Extract runtime environment CLI commands
